src/load_generator/pyproject.toml
[project]
name = "load-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.0",
    "numpy>=2.2.5",
    "prometheus-client>=0.21.1",
    "pyyaml>=6.0.2",
    "watchdog>=6.0.0",
]


[tool.ruff]
line-length = 120
